pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "qtzmq"
version = "0.1.3"
description = "Qt-friendly ZeroMQ bindings for PySide and PyQt applications"
readme = "README.md"
requires-python = ">=3.9"

authors = [
    { name = "Roberto Herrera", email = "robertodherreram@gmail.com" }
]

license = { text = "MIT" }

keywords = [
    "qt",
    "pyside",
    "pyqt",
    "zeromq",
    "zmq",
    "pyzmq",
    "qt signals",
    "pubsub",
    "qt gui",
    "instrument control"
]

dependencies = [
    "pyzmq"
]

[project.optional-dependencies]
qt = [
    "PySide6"
]

[project.urls]
Homepage = "https://github.com/Robertoherrera0/qtzmq"
Source = "https://github.com/Robertoherrera0/qtzmq"
Issues = "https://github.com/Robertoherrera0/qtzmq/issues"